qrcode updates, navigator push update
