+++ rpg_go/lib/components/floating_menu_buttons.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:popover/popover.dart';
+import 'package:rpg_go/pages/qr_scan.dart';
 //import 'package:rpg_go/pages/master_room.dart';
-import 'package:rpg_go/pages/room_creation.dart';
+import 'package:rpg_go/pages/table_creation.dart';
+import 'package:rpg_go/pages/select_sheet.dart';
 
 class FloatMasterButton extends StatefulWidget {
   const FloatMasterButton({super.key});
@@ -22,9 +24,9 @@ class _FloatMasterButtonState extends State<FloatMasterButton> {
           height: 150,
           context: context,
           backgroundColor: mainColor,
-          bodyBuilder: (context) => Column(
+          bodyBuilder: (context) => const Column(
                 children: [
-                  const Text(
+                  Text(
                     "EU SOU:",
                     style: TextStyle(
                         fontSize: 20, fontFamily: 'Revol', color: Colors.white),
@@ -32,62 +34,11 @@ class _FloatMasterButtonState extends State<FloatMasterButton> {
                   Row(
                     mainAxisAlignment: MainAxisAlignment.spaceAround,
                     children: [
-                      ElevatedButton(
-                          style: ElevatedButton.styleFrom(
-                              backgroundColor: Colors.deepPurple,
-                              padding: const EdgeInsets.all(20),
-                              shape: RoundedRectangleBorder(
-                                  borderRadius: BorderRadius.circular(12))),
-                          onPressed: () {
-                            Navigator.push(
-                                context,
-                                MaterialPageRoute(
-                                    builder: (context) => const RoomCreation()));
-                          },
-                          child: const Column(
-                            children: [
-                              Icon(
-                                Icons.edit,
-                                color: Colors.white,
-                              ),
-                              Text(
-                                "Mestre",
-                                style: TextStyle(
-                                    fontSize: 20,
-                                    color: Colors.white,
-                                    fontFamily: 'Revol'),
-                              ),
-                            ],
-                          )),
-                      const SizedBox(
+                      MasterButton(),
+                      SizedBox(
                         width: 50,
                       ),
-                      ElevatedButton(
-                        onPressed: () {},
-                        style: ElevatedButton.styleFrom(
-                            backgroundColor: Colors.blueGrey,
-                            padding: const EdgeInsets.all(20),
-                            shape: RoundedRectangleBorder(
-                                borderRadius: BorderRadius.circular(12))),
-                        child: const Column(
-                          children: [
-                            Icon(
-                              Icons.qr_code,
-                              color: Colors.white,
-                            ),
-                            FittedBox(
-                              fit: BoxFit.scaleDown,
-                              child: Text(
-                                "Jogador",
-                                style: TextStyle(
-                                    fontSize: 20,
-                                    color: Colors.white,
-                                    fontFamily: 'Revol'),
-                              ),
-                            ),
-                          ],
-                        ),
-                      )
+                      PlayerButton()
                     ],
                   ),
                 ],
@@ -101,5 +52,75 @@ class _FloatMasterButtonState extends State<FloatMasterButton> {
         color: Colors.white,
       ),
     );
+  }
+}
+
+class PlayerButton extends StatelessWidget {
+  const PlayerButton({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return ElevatedButton(
+      onPressed: () {
+        Navigator.push(
+            context, MaterialPageRoute(builder: (context) => SelectSheet()));
+      },
+      style: ElevatedButton.styleFrom(
+          backgroundColor: Colors.blueGrey,
+          padding: const EdgeInsets.all(20),
+          shape:
+              RoundedRectangleBorder(borderRadius: BorderRadius.circular(12))),
+      child: const Column(
+        children: [
+          Icon(
+            Icons.qr_code,
+            color: Colors.white,
+          ),
+          FittedBox(
+            fit: BoxFit.scaleDown,
+            child: Text(
+              "Jogador",
+              style: TextStyle(
+                  fontSize: 20, color: Colors.white, fontFamily: 'Revol'),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class MasterButton extends StatelessWidget {
+  const MasterButton({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return ElevatedButton(
+        style: ElevatedButton.styleFrom(
+            backgroundColor: Colors.deepPurple,
+            padding: const EdgeInsets.all(20),
+            shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(12))),
+        onPressed: () {
+          Navigator.push(context,
+              MaterialPageRoute(builder: (context) => const TableCreation()));
+        },
+        child: const Column(
+          children: [
+            Icon(
+              Icons.edit,
+              color: Colors.white,
+            ),
+            Text(
+              "Mestre",
+              style: TextStyle(
+                  fontSize: 20, color: Colors.white, fontFamily: 'Revol'),
+            ),
+          ],
+        ));
   }
 }
