+++ rpg_go/lib/main.dart
@@ -3,6 +3,7 @@ import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:rpg_go/pages/home_revival.dart';
 import 'package:rpg_go/pages/login_page.dart';
 import 'package:rpg_go/pages/master_room.dart';
+import 'package:rpg_go/pages/profile.dart';
 import 'package:rpg_go/pages/qr_scan.dart';
 
 //import 'pages/attack_page.dart';
@@ -10,21 +11,24 @@ import 'package:rpg_go/pages/qr_scan.dart';
 
 Future main() async {
   await dotenv.load(fileName: "lib/.env");
-  runApp(
-    const MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({super.key}); 
+  const MyApp({super.key});
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
-      home: const MasterRoom(),
+      home: LoginPage(),
       routes: {
         '/home_revival': (context) => HomeRevival(),
-        '/master_room': (context) =>  const MasterRoom(),
+        '/master_room': (context) => MasterRoom(
+              name: "",
+              id: 0,
+            ),
+        '/profile': (context) => const Profile()
       },
     );
   }
