--- rpg_go/lib/models/Table.dart
@@ -1,31 +1,0 @@
-import 'package:rpg_go/models/Room.dart';
-
-class Table {
-  int id;
-  String name;
-  List<Room>? rooms =[];
-  int masterId;
-
-  Table({
-    required this.id,
-    required this.name,
-    this.rooms,
-    required this.masterId,
-  });
-
-Table.fromJson(Map<String, dynamic> json)
-      : id = json['id'],
-        name = json['name'],
-        rooms = json['rooms'] != null
-            ? List<dynamic>.from(json['rooms']).map((i) => Room.fromJson(i)).toList()
-            : List.empty(),
-        masterId = json['masterId'];
-
-  Map<String, dynamic> toJson() => {
-        'id': id,
-        'name': name,
-        'rooms': rooms,
-        'masterId': masterId,
-      };
-  
-}
